gitea: fix handling when `fetchFromGitea` falls back to `fetchgit`
`fetchFromGitea` just re-uses `fetchFromGitHub`.
The only difference is making `domain` a required attribute instead of
defaulting to `"github.com"`.

As such, it falls back to `fetchgit`, such like `fetchFromGitHub`, if
any of the following is true:

 - fetchSubmodules
 - leaveDotGit
 - deepClose
 - sparseCheckout != []
 - forceFetchGit

When falling back to `fetchgit`, the `src.url` will end with `".git"`,
which we then need to strip again from our parsed `repo` variable.

`fetchFromGitHub` has the same logic to trim `".git"` unconditionally.