github: Revert use of API to determine latest version
This behavior does not work for any repository that has used GitHub
releases in the past but only uses tags for newer versions. The original
motivation for the change was to prevent pre-releases from being
selected as the latest version, but the `--version-regex` flag exists
precisely for the purpose of selecting only intended versions (and not
unrelated versions, pre-releases, dev tags, etc).

As such, the change provided no benefit to the purported motivation, but
it did break updates for quite a few packages and has resulted in
downgrades and misversioning, among other things.

A possible attempt to include the GitHub API could be used to
specifically exclude tags associated with pre-releases in the GitHub
API, but otherwise continue using tags as the "source of truth" for the
latest version. This change does not attempt that as it is only intended
to make nix-update work correctly as it did before the change.

This reverts commits bf33dc45ce26f36233fd58790f90dc8710ca7198 and
696efb62a9df3223a255b15e005e4d0815130a0c.