Match last "got" in stderr to avoid false positives
Due to the changes introduced in #349 the hash regex matches the output
of pnpm when downloading the https://www.npmjs.com/package/got library.
This causes the tool to incorrectly update the hash to the "got"
version. To fix this I simply changed the for loop to look for the last
regex match in the stderr instead of last, since the nix hash mismatch
message is usually printed last.

Signed-off-by: Marcin Serwin <marcin@serwin.dev>