Match last "got" in stderr to avoid false positives
Due to the changes introduced in #349 the hash regex matches the output
of pnpm when downloading the https://www.npmjs.com/package/got library.
This causes the tool to incorrectly update the hash to the "got"
version. To fix this the loop now starts searching from the last line,
since the nix hash mismatch message is usually printed last.

Signed-off-by: Marcin Serwin <marcin@serwin.dev>